♻️ Move adjust edit pages under adjust-edit dir
