+++ src/pages/adjust-edit/AdjSubjectCriteria/adj-subject-criteria.module.css
@@ -1,4 +1,4 @@
-@import url("../../styles/global.css");
+@import url('../../../styles/global.css');
 
 .titleWrapper {
   display: flex;
@@ -125,7 +125,7 @@
 .frame2 {
   display: flex;
   flex-direction: row;
-  flex-wrap: wrap;   
+  flex-wrap: wrap;
   align-items: flex-start;
   justify-content: flex-start;
   gap: 8px;
@@ -136,7 +136,7 @@
 .frame3 {
   display: flex;
   flex-direction: column;
-  flex-wrap: wrap;   
+  flex-wrap: wrap;
   align-items: flex-start;
   justify-content: flex-start;
   gap: 8px;
@@ -166,7 +166,7 @@
   transform: rotate(0deg) scale(1, 1);
 }
 
-.errorMessage{
+.errorMessage {
   font: var(--font-caption-1r-pc);
   color: var(--color-system-caution);
 }
@@ -180,7 +180,7 @@
 .groupRow {
   display: flex;
   flex-wrap: wrap; /* 너무 좁으면 자동 줄바꿈 */
-  gap: 4px;        /* 각 스위치 사이 간격 */
+  gap: 4px; /* 각 스위치 사이 간격 */
   align-items: center;
 }
