+++ src/pages/adjust-edit/HighOrganization/high-organization-page.module.css
@@ -1,5 +1,5 @@
-@import url('../../styles/global.css');
-@import url('../../styles/table.css');
+@import url('../../../styles/global.css');
+@import url('../../../styles/table.css');
 
 .section {
   display: flex;
