+++ src/pages/adjust-edit/PaybandApply/payband-apply-page.module.css
@@ -1,5 +1,5 @@
-@import url('../../styles/global.css');
-@import url('../../styles/table.css');
+@import url('../../../styles/global.css');
+@import url('../../../styles/table.css');
 
 .payband-apply-area {
   display: flex;
