+++ src/pages/adjust-edit/PaybandConfig/payband-config-page.module.css
@@ -1,5 +1,5 @@
-@import url('../../styles/global.css');
-@import url('../../styles/table.css');
+@import url('../../../styles/global.css');
+@import url('../../../styles/table.css');
 .page {
   width: 100%;
   display: flex;
