+++ src/pages/adjust-edit/ResultPage/Card.jsx
