+++ src/pages/adjust-edit/ResultPage/result-page.module.css
@@ -1,5 +1,5 @@
-@import url('../../styles/global.css');
-@import url('../../styles/table.css');
+@import url('../../../styles/global.css');
+@import url('../../../styles/table.css');
 
 .page-nation-area {
   width: 100%;
